Moving AirflowClient to clients module
